feat: delete user

+++ apps/raise/src/qa-helper/commands/purge.user.command.ts/purge.user.command.ts
@@ -269,7 +269,10 @@ export class PurgeUserCommandHandler
         },
       );
 
-      if (deleted_file_manager_ids.length === res.deletedFileCount) {
+      if (
+        deleted_file_manager_ids.length > 0 &&
+        deleted_file_manager_ids.length === res.deletedFileCount
+      ) {
         for (const url of deleted_file_manager_urls) {
           await this.bunnyService.deleteMedia(url, true);
         }
